Add Member Pro reports (#12)
* Add Member Pro reports

* Add mod checks

* fix commands length

* merge report member

* fix perms

+++ utils/views/menus/report.py
@@ -11,6 +11,7 @@ from utils import GIRContext, cfg
 from utils.framework import gatekeeper
 from utils.mod import ban, mute, unmute, warn
 from utils.views.modals.prompt import GenericDescriptionModal
+from utils.framework.checks import PermissionsFailure
 
 from .report_action import ModAction, ReportActionReason
 
@@ -70,6 +71,31 @@ async def manual_report(mod: discord.Member, target: Union[discord.Message, disc
         view = ReportActions(target)
 
     embed = prepare_embed(target, title="A moderator reported a member")
+    await channel.send(ping_string, embed=embed, view=view, allowed_mentions=discord.AllowedMentions(everyone=False, users=True, roles=True))
+
+async def mempro_report(genius: discord.Member, target: Union[discord.Message, discord.Member] = None):
+    """Deals with a report
+
+    Parameters
+    ----------
+    bot : discord.Client
+        "Bot object"
+    message : discord.Message
+        "Filtered message"
+    genius : discord.Member
+        "The genius that started this report
+
+    """
+    db_guild = guild_service.get_guild()
+    channel = target.guild.get_channel(db_guild.channel_mempro_reports)
+
+    ping_string = f"{genius.mention} reported a member"
+    if isinstance(target, discord.Message):
+        view = ReportActions(target.author)
+    else:
+        view = ReportActions(target)
+
+    embed = prepare_embed(target, title="A user reported a member")
     await channel.send(ping_string, embed=embed, view=view, allowed_mentions=discord.AllowedMentions(everyone=False, users=True, roles=True))
 
 
@@ -235,11 +261,18 @@ class ReportActions(ui.View):
 
     @ui.button(emoji="✅", label="Dismiss", style=discord.ButtonStyle.primary)
     async def dismiss(self, interaction: discord.Interaction, _: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
         await interaction.message.delete()
         self.stop()
 
     @ui.button(emoji="⚠️", label="Warn", style=discord.ButtonStyle.primary)
     async def warn(self, interaction: discord.Interaction, _: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
+
         view = ReportActionReason(
             target_member=self.target_member, mod=interaction.user, mod_action=ModAction.WARN)
         await interaction.response.send_message(embed=discord.Embed(description=f"{interaction.user.mention}, choose a warn reason for {self.target_member.mention}.", color=discord.Color.blurple()), view=view)
@@ -252,6 +285,10 @@ class ReportActions(ui.View):
 
     @ui.button(emoji="❌", label="Ban", style=discord.ButtonStyle.primary)
     async def ban(self, interaction: discord.Interaction, _: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
+
         view = ReportActionReason(
             target_member=self.target_member, mod=interaction.user, mod_action=ModAction.BAN)
         await interaction.response.send_message(embed=discord.Embed(description=f"{interaction.user.mention}, choose a ban reason for {self.target_member.mention}.", color=discord.Color.blurple()), view=view)
@@ -264,17 +301,29 @@ class ReportActions(ui.View):
 
     @ui.button(emoji="🆔", label="Post ID", style=discord.ButtonStyle.primary)
     async def id(self, interaction: discord.Interaction, _: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
+
         await interaction.response.send_message(self.target_member.id)
         await asyncio.sleep(10)
         await interaction.delete_original_message()
 
     @ui.button(emoji="🧹", label="Clean up", style=discord.ButtonStyle.primary)
     async def purge(self, interaction: discord.Interaction, button: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
+
         await interaction.channel.purge(limit=100)
         self.stop()
 
     @ui.button(emoji="🔎", label="Claim report", style=discord.ButtonStyle.primary)
     async def claim(self, interaction: discord.Interaction, button: ui.Button):
+        if not gatekeeper.has(interaction.guild, interaction.user, 5):
+            raise PermissionsFailure(
+                "You do not have permission to use this command.")
+
         report_embed = interaction.message.embeds[0]
         if "(claimed)" in report_embed.title:
             ctx = GIRContext(interaction)
